loadgen/pyproject.toml
[project]
name = "sql-ces-loadgen"
version = "0.1.0"
description = "Synthetic load generator for SQL CES POC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyodbc>=5.1.0",
    "yoyo-migrations>=8.2.0",
    "faker>=24.0.0",
    "structlog>=24.1.0",
]

[project.scripts]
loadgen = "loadgen.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loadgen"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]
